Combined Tracker and Preparer into Handler #46
