+++ src/hawkengine/src/main/java/net/hawkengine/core/materialupdater/IMaterialHandlerService.java
@@ -1,7 +1,10 @@
 package net.hawkengine.core.materialupdater;
 
+import net.hawkengine.model.Pipeline;
 import net.hawkengine.model.PipelineDefinition;
 
-public interface IMaterialTrackerService {
+public interface IMaterialHandlerService {
     String checkPipelineForTriggerMaterials(PipelineDefinition pipelineDefinition);
+
+    Pipeline updatePipelineMaterials(Pipeline pipeline);
 }
