--- src/hawkengine/src/main/java/net/hawkengine/core/materialupdater/IMaterialPreparerService.java
@@ -1,7 +1,0 @@
-package net.hawkengine.core.materialupdater;
-
-import net.hawkengine.model.Pipeline;
-
-public interface IMaterialPreparerService {
-    Pipeline updatePipelineMaterials(Pipeline pipeline);
-}
